ajout findAll

+++ app/src/main/java/com/w32/tp2_samuel_emma/repository/Repository.java
@@ -9,5 +9,5 @@ public interface Repository<T> {
     T findLast();
     boolean update(T dataObject);
     boolean delete(T dataObject);
-    boolean delete(int id);
+    boolean delete(long id);
 }
